feat: compiler: bin-expr almost done
